requirements.txt
google-adk
litellm
requests 